+++ src/RPlatform.h
@@ -39,7 +39,7 @@ public:
   void setRunSettings(direction dirL, direction dirR, int powerL, int powerR); // настройка всех параметров двигателей
   void run();                                                                  // запуск двигателей
   void move(int leftPower, int rightPower);
-  void move(int leftPower, int rightPower, int time);
+  void move(int leftPower, int rightPower, float time);
   void move(int leftPower, int rightPower, int lAngle, int rAngle);
   void runTime(float seconds);           // запуск двигателей на N секунд
   void runSteps(int steps);              // запуск двигателей на N шагов
